requirements.txt
fastapi
uvicorn[standard]
pymongo
python-dotenv
cloudinary
celery
redis
ffmpeg-python
pydantic
python-multipart